@@ -1,14 +1,14 @@
-import 'package:hive/hive.dart';
-
+import 'package:isar/isar.dart';
 import 'exercise.dart';
 
 part 'workout.g.dart';
-@HiveType(typeId: 77, adapterName: 'WorkoutAdapter')
-class Workout {
-  @HiveField(0)
-  String name;
-  @HiveField(1)
-  final List<Exercise> exercises;
 
-  Workout({required this.name, required this.exercises});
+@Collection()
+class Workout {
+  Id id = Isar.autoIncrement;
+
+  late String name;
+  String? description;
+
+  final exercises = IsarLinks<Exercise>();
 }
